Various fixes for vendor and more tests, and config caching (#71)
* dependency vendor omitempty

* wmake format

* flattening tests vendor

* Cleaning up some def/flag code

* wmake format

* using headers to reduce packageData size

* tests for singleton

* files for windows dev

* fixing path errors

* wmake format

* test for singleton

* Fixing flag issues and update

* made Node fields public

* pass Node pointer to resolve remote

* extra information added to Node for making build easier

* fixed a bug inside resolve

* Some code cleanup and adding PackageCache

+++ tests/symlinks.sh
@@ -1,5 +1,13 @@
 #!/bin/bash
 
+set -e
+
+# Check that working directory contains script
+if [ ! -f $(pwd)/`basename "$0"` ]; then
+    printf "ERROR: `basename "$0"` must be run from its directory\n"
+    exit 1
+fi
+
 mkdir -p $(pwd)/project-pkg/pkg-list/vendor
 rm -f $(pwd)/project-pkg/pkg-list/vendor/pkg-malloc
 ln -s -f $(pwd)/project-pkg/pkg-malloc $(pwd)/project-pkg/pkg-list/vendor/pkg-malloc
@@ -7,9 +15,25 @@ ln -s -f $(pwd)/project-pkg/pkg-malloc $(pwd)/project-pkg/pkg-list/vendor/pkg-ma
 mkdir -p $(pwd)/project-app/app-avr/vendor
 rm -f $(pwd)/project-app/app-avr/vendor/pkg-uart
 rm -f $(pwd)/project-app/app-avr/vendor/pkg-list
+rm -f $(pwd)/project-app/app-avr/vendor/pkg-malloc
 ln -s -f $(pwd)/project-pkg/pkg-uart $(pwd)/project-app/app-avr/vendor/pkg-uart
 ln -s -f $(pwd)/project-pkg/pkg-list $(pwd)/project-app/app-avr/vendor/pkg-list
+ln -s -f $(pwd)/project-pkg/pkg-malloc $(pwd)/project-app/app-avr/vendor/pkg-malloc
 
 mkdir -p $(pwd)/project-pkg/pkg-trace/vendor
 rm -f $(pwd)/project-pkg/pkg-trace/vendor/pkg-uart
 ln -s -f $(pwd)/project-pkg/pkg-uart $(pwd)/project-pkg/pkg-trace/vendor/pkg-uart
+
+mkdir -p $(pwd)/project-app/app-alloc/vendor
+mkdir -p $(pwd)/project-app/app-alloc/vendor/alloc-one/vendor
+mkdir -p $(pwd)/project-app/app-alloc/vendor/alloc-two/vendor
+rm -f $(pwd)/project-app/app-alloc/vendor/pkg-malloc
+rm -f $(pwd)/project-app/app-alloc/vendor/alloc-one/vendor/pkg-malloc
+rm -f $(pwd)/project-app/app-alloc/vendor/alloc-two/vendor/pkg-malloc
+ln -s -f $(pwd)/project-pkg/pkg-malloc $(pwd)/project-app/app-alloc/vendor/pkg-malloc
+ln -s -f $(pwd)/project-pkg/pkg-malloc $(pwd)/project-app/app-alloc/vendor/alloc-one/vendor/pkg-malloc
+ln -s -f $(pwd)/project-pkg/pkg-malloc $(pwd)/project-app/app-alloc/vendor/alloc-two/vendor/pkg-malloc
+
+mkdir -p $(pwd)/project-pkg/pkg-singleton/vendor
+rm -f $(pwd)/project-pkg/pkg-singleton/vendor/pkg-malloc
+ln -s -f $(pwd)/project-pkg/pkg-malloc $(pwd)/project-pkg/pkg-singleton/vendor/pkg-malloc
